refactor(afternote): 데이터 매퍼 클래스 분리 및 구조 개선
- `AfternoteMapper.kt`에 집중되어 있던 매퍼 함수들을 역할에 따라 분리
- `PlaylistInputMapper.kt`: 플레이리스트 입력 관련 DTO 변환 로직 추가
- `ResponseMapper.kt`: API 응답값에 대한 도메인 모델 변환 로직 이동 및 가독성 개선
- `AfternoteInputMapper.kt`: 중복된 DTO 변환 로직 제거
- `formatDateFromServer` 함수 접근 제어자를 `internal`로 변경하여 공유 가능하도록 수정
