+++ app/src/main/java/com/kuit/afternote/feature/afternote/data/mapper/AfternoteInputMapper.kt
@@ -2,38 +2,17 @@ package com.kuit.afternote.feature.afternote.data.mapper
 
 import com.kuit.afternote.feature.afternote.data.dto.AfternoteCredentials
 import com.kuit.afternote.feature.afternote.data.dto.AfternoteMemorialVideo
-import com.kuit.afternote.feature.afternote.data.dto.AfternotePlaylist
 import com.kuit.afternote.feature.afternote.data.dto.AfternoteReceiverRef
-import com.kuit.afternote.feature.afternote.data.dto.AfternoteSong
 import com.kuit.afternote.feature.afternote.data.dto.request.AfternoteUpdateRequest
 import com.kuit.afternote.feature.afternote.domain.model.CredentialsInput
 import com.kuit.afternote.feature.afternote.domain.model.ReceiverRefInput
 import com.kuit.afternote.feature.afternote.domain.model.UpdateRequestInput
 import com.kuit.afternote.feature.afternote.domain.model.playlist.MemorialVideoInput
-import com.kuit.afternote.feature.afternote.domain.model.playlist.PlaylistInput
-import com.kuit.afternote.feature.afternote.domain.model.playlist.SongInput
-
-fun PlaylistInput.toDto() =
-    AfternotePlaylist(
-        profilePhoto = profilePhoto,
-        atmosphere = atmosphere,
-        memorialPhotoUrl = memorialPhotoUrl,
-        songs = songs.map { it.toDto() },
-        memorialVideo = memorialVideo?.toDto(),
-    )
 
 fun MemorialVideoInput.toDto() =
     AfternoteMemorialVideo(
         videoUrl = videoUrl,
         thumbnailUrl = thumbnailUrl,
-    )
-
-fun SongInput.toDto() =
-    AfternoteSong(
-        id = id,
-        title = title,
-        artist = artist,
-        coverUrl = coverUrl,
     )
 
 fun UpdateRequestInput.toDto() =
